Update to version 0.4.1

+++ src/Commands/2_admin/prune.js
@@ -1,7 +1,7 @@
 const Command = require('../../Structures/Command.js');
 
-const { Permissions } = require('discord.js');
-const { bot: { prefix, ownerID }, emoji: { success, warning }, response: { missingArguments, invalidPermissions, invalidNumber } } = require('../../../config/config.json');
+const { PermissionsBitField } = require('discord.js');
+const { bot: { prefix, ownerID }, emoji: { success, warning, error }, response: { missingArguments, invalidPermissions, invalidNumber } } = require('../../../config/config.json');
 
 module.exports = new Command({
 	name: 'prune',
@@ -9,33 +9,39 @@ module.exports = new Command({
     syntax: 'prune <amount>',
 	description: 'Deletes the amount of messages send by the bot and the commands used to invoke the bot. Requires Manage Messages permission.',
 	async run(message, args, client) {
-		if (!message.member.permissions.has(Permissions.FLAGS.MANAGE_MESSAGES) && message.author.id != ownerID) return message.channel.send(`${warning} ${invalidPermissions} (Manage Messages)`);
+		if (!message.member.permissions.has(PermissionsBitField.Flags.ManageMessages) && message.author.id != ownerID) return message.channel.send(`${warning} ${invalidPermissions} (Manage Messages)`);
         if (!args[0]) return message.channel.send(`${warning} ${missingArguments}`);
         if (isNaN(args[0])) return message.channel.send(`${warning} ${invalidNumber}`);
         if (args[0] > 100 || args[0] < 1) return message.channel.send(`${warning} Outside of number range!`);
+
+        const response = await message.channel.send(`${success} Deleting up to ${args[0]} messages. This may take a while...`)
 
         const result = await message.channel.messages.fetch({limit: 100});
 
         const resultArray = new Array();
         result.each(message => resultArray.push(message));
 
+        const messagesToBeDeleted = new Array();
+
         let messagesDeleted = 0;
-        for (let i = 0; i < resultArray.length; i++) {
-            if (resultArray[i].author.id == client.user.id) {
-                resultArray[i].delete();
+        for (let i = 2; i < resultArray.length; i++) {
+            if (resultArray[i] && resultArray[i].author.id == client.user.id && resultArray[i].createdTimestamp < (Date.now() - 8000)) { //older than 8 seconds for reactions
+                messagesToBeDeleted.push(resultArray[i]);
                 messagesDeleted++;
-                if (resultArray[i+1].content.startsWith(prefix)) {
-                    resultArray[i+1].delete();
+                if (resultArray[i+1] && resultArray[i+1].content.startsWith(prefix)) {
+                    messagesToBeDeleted.push(resultArray[i+1]);
                     messagesDeleted++;
                 }
             }
             if (messagesDeleted >= args[0]) break;
         }
 
-        const response = await message.channel.send(`${success} Deleting ${args[0]} messages`)
-        setTimeout(() => {
-            response.delete();
-            message.delete();
+        message.channel.bulkDelete(messagesToBeDeleted, false);
+
+        setTimeout( async () => {
+            const fetched = await response.channel.messages.fetch({ limit: 2, cache: false, around: response.id });
+            if (fetched.has(response.id)) response.delete();
+            if (fetched.has(message.id)) message.delete();
         }, 3750);
 	}
 });
